Remove registry
