+++ Cargo.toml
@@ -45,7 +45,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1.0"
 sha2 = "0.7"
 shell-escape = "0.1"
-simsearch = "0.1"
+simsearch = "0.2"
 slog = { version = "2", features = ["max_level_trace", "release_max_level_warn"] }
 slog-async = "2"
 slog-term = "2"
