+++ src/lib/cli/mod.rs
@@ -1,5 +1,5 @@
 //! Handlers for all of the command-line actions of the binary.
 
 pub mod build;
+pub mod index;
 pub mod new;
-pub mod registry;
