+++ src/lib/remote/mod.rs
@@ -1,5 +1,4 @@
 mod index;
-mod registry;
 pub mod resolution;
 
-pub use self::{index::*, registry::*};
+pub use self::index::*;
